+++ spec/helper.rb
@@ -1,15 +1,13 @@
 $TESTING = true
 
-if RUBY_VERSION >= "1.9"
-  require "simplecov"
-  require "coveralls"
+require "simplecov"
+require "coveralls"
 
-  SimpleCov.formatters = [SimpleCov::Formatter::HTMLFormatter, Coveralls::SimpleCov::Formatter]
+SimpleCov.formatters = [SimpleCov::Formatter::HTMLFormatter, Coveralls::SimpleCov::Formatter]
 
-  SimpleCov.start do
-    add_filter "/spec"
-    minimum_coverage(90)
-  end
+SimpleCov.start do
+  add_filter "/spec"
+  minimum_coverage(90)
 end
 
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), "..", "lib"))
@@ -25,10 +23,10 @@ require "webmock/rspec"
 WebMock.disable_net_connect!(:allow => "coveralls.io")
 
 # Set shell to basic
+ENV["THOR_COLUMNS"] = "10000"
 $0 = "thor"
 $thor_runner = true
 ARGV.clear
-Encoding.default_external = Encoding::UTF_8 if RUBY_VERSION > '1.8.7'
 Thor::Base.shell = Thor::Shell::Basic
 
 # Load fixtures
@@ -77,6 +75,13 @@ RSpec.configure do |config|
     yield
   ensure
     $VERBOSE = old_verbose
+  end
+
+  # true if running on windows, used for conditional spec skips
+  #
+  # @return [TrueClass/FalseClass]
+  def windows?
+    Gem.win_platform?
   end
 
   alias silence capture