+++ spec/parser/options_spec.rb
@@ -6,7 +6,6 @@ describe Thor::Options do
     opts.each do |key, value|
       opts[key] = Thor::Option.parse(key, value) unless value.is_a?(Thor::Option)
     end
-
     @opt = Thor::Options.new(opts, defaults, stop_on_unknown)
   end
 
@@ -113,7 +112,11 @@ describe Thor::Options do
     it "raises an error for unknown switches" do
       create :foo => "baz", :bar => :required
       parse("--bar", "baz", "--baz", "unknown")
-      expect { check_unknown! }.to raise_error(Thor::UnknownArgumentError, "Unknown switches '--baz'")
+
+      expected = "Unknown switches \"--baz\""
+      expected << "\nDid you mean?  \"--bar\"" if Thor::Correctable
+
+      expect { check_unknown! }.to raise_error(Thor::UnknownArgumentError, expected)
     end
 
     it "skips leading non-switches" do
@@ -298,6 +301,13 @@ describe Thor::Options do
         expect { parse("--fruit", "orange") }.to raise_error(Thor::MalformattedArgumentError,
             "Expected '--fruit' to be one of #{enum.join(', ')}; got orange")
       end
+
+      it "allows multiple values if repeatable is specified" do
+        create :foo => Thor::Option.new("foo", :type => :string, :repeatable => true)
+
+        expect(parse("--foo=bar", "--foo", "12")["foo"]).to eq(["bar", "12"])
+        expect(parse("--foo", "13", "--foo", "14")["foo"]).to eq(["bar", "12", "13", "14"])
+      end
     end
 
     describe "with :boolean type" do
@@ -327,6 +337,10 @@ describe Thor::Options do
         expect(parse("--skip-foo")["foo"]).to eq(false)
       end
 
+      it "accepts --[skip-]opt variant, setting false for value, even if there's a trailing non-switch" do
+        expect(parse("--skip-foo", "asdf")["foo"]).to eq(false)
+      end
+
       it "will prefer 'no-opt' variant over inverting 'opt' if explicitly set" do
         create "--no-foo" => true
         expect(parse("--no-foo")["no-foo"]).to eq(true)
@@ -337,6 +351,19 @@ describe Thor::Options do
         expect(parse("--skip-foo")["skip-foo"]).to eq(true)
       end
 
+      it "will prefer 'skip-opt' variant over inverting 'opt' if explicitly set, even if there's a trailing non-switch" do
+        create "--skip-foo" => true
+        expect(parse("--skip-foo", "asdf")["skip-foo"]).to eq(true)
+      end
+
+      it "will prefer 'skip-opt' variant over inverting 'opt' if explicitly set, and given a value" do
+        create "--skip-foo" => true
+        expect(parse("--skip-foo=f")["skip-foo"]).to eq(false)
+        expect(parse("--skip-foo=false")["skip-foo"]).to eq(false)
+        expect(parse("--skip-foo=t")["skip-foo"]).to eq(true)
+        expect(parse("--skip-foo=true")["skip-foo"]).to eq(true)
+      end
+
       it "accepts inputs in the human name format" do
         create :foo_bar => :boolean
         expect(parse("--foo-bar")["foo_bar"]).to eq(true)
@@ -345,21 +372,23 @@ describe Thor::Options do
       end
 
       it "doesn't eat the next part of the param" do
-        create :foo => :boolean
         expect(parse("--foo", "bar")).to eq("foo" => true)
         expect(@opt.remaining).to eq(%w(bar))
       end
 
       it "doesn't eat the next part of the param with 'no-opt' variant" do
-        create :foo => :boolean
         expect(parse("--no-foo", "bar")).to eq("foo" => false)
         expect(@opt.remaining).to eq(%w(bar))
       end
 
       it "doesn't eat the next part of the param with 'skip-opt' variant" do
-        create :foo => :boolean
         expect(parse("--skip-foo", "bar")).to eq("foo" => false)
         expect(@opt.remaining).to eq(%w(bar))
+      end
+
+      it "allows multiple values if repeatable is specified" do
+        create :verbose => Thor::Option.new("verbose", :type => :boolean, :aliases => '-v', :repeatable => true)
+        expect(parse("-v", "-v", "-v")["verbose"].count).to eq(3)
       end
     end
 
@@ -383,6 +412,11 @@ describe Thor::Options do
       it "must not allow the same hash key to be specified multiple times" do
         expect { parse("--attributes", "name:string", "name:integer") }.to raise_error(Thor::MalformattedArgumentError, "You can't specify 'name' more than once in option '--attributes'; got name:string and name:integer")
       end
+
+      it "allows multiple values if repeatable is specified" do
+        create :attributes => Thor::Option.new("attributes", :type => :hash, :repeatable => true)
+        expect(parse("--attributes", "name:one", "foo:1", "--attributes", "name:two", "bar:2")["attributes"]).to eq({"name"=>"two", "foo"=>"1", "bar" => "2"})
+      end
     end
 
     describe "with :array type" do
@@ -400,6 +434,11 @@ describe Thor::Options do
 
       it "must not mix values with other switches" do
         expect(parse("--attributes", "a", "b", "c", "--baz", "cool")["attributes"]).to eq(%w(a b c))
+      end
+
+      it "allows multiple values if repeatable is specified" do
+        create :attributes => Thor::Option.new("attributes", :type => :array, :repeatable => true)
+        expect(parse("--attributes", "1", "2", "--attributes", "3", "4")["attributes"]).to eq([["1", "2"], ["3", "4"]])
       end
     end
 
@@ -426,6 +465,11 @@ describe Thor::Options do
         create :limit => Thor::Option.new("limit", :type => :numeric, :enum => enum)
         expect { parse("--limit", "3") }.to raise_error(Thor::MalformattedArgumentError,
                                                         "Expected '--limit' to be one of #{enum.join(', ')}; got 3")
+      end
+
+      it "allows multiple values if repeatable is specified" do
+        create :run => Thor::Option.new("run", :type => :numeric, :repeatable => true)
+        expect(parse("--run", "1", "--run", "2")["run"]).to eq([1, 2])
       end
     end
   end